♻️ refactor: 모듈 엔트리 포인트로 컴포넌트 경로 수정
