+++ pandamarket/src/App.js
@@ -1,5 +1,4 @@
-import Footer from "./components/Footer";
-import Header from "./components/Header";
+import {Header, Footer} from './components'
 import Main from "./pages/Main";
 import Items from "./pages/Items";
 import { Routes, Route, BrowserRouter } from "react-router-dom";
