+++ pandamarket/src/components/index.js
@@ -1,0 +1,6 @@
+export { default as BestProductList } from "./BestProductList";
+export { default as ProductList } from "./ProductList";
+export { default as Footer } from "./Footer";
+export { default as Header } from "./Header";
+export { default as Pagination } from "./Pagination";
+
